Update how we generate .pc file
Update how we generate pkg-config files.
Use the library object as a base for the .pc file, this gives us some
information for the .pc file for free.
Add URL.
Update include directory since we install headers in a subdir (this was
broken before)
Require a newver version of meson that includes support for this way of
generating .pc files.