First pass at using the Symbol map
First pass at using the Symbol map to hide symbols.
This is a little clunky in meson, since there's no built in support for
this yet.  What we do is we define the path to the version map, and the
linker flags to use it, and then test if the compiler is OK with this.
If it is, then all is good, otherwise, we abort and complain.
The linker flags are also added to to library build step.